Main

Description

Custom XMonad configuration.

Documentation

bg :: String

Default colors.

fg :: String

Default colors.

blue :: String

Default colors.

green :: String

Default colors.

red :: String

Default colors.

orange :: [Char]

workspaces' :: [WorkspaceId]

List of workspace names.

terminal' :: String

Default terminal.

borderWidth' :: Dimension

Xmonad border width.

modMask' :: KeyMask

Use Super_L as Mod key.

normalBorderColor' :: String

Xmonad border colors.

focusedBorderColor' :: String

Xmonad border colors.

spacing' :: Int

Space around windows.

layout' :: ModifiedLayout Rename (ModifiedLayout (ConfigurableBorder Ambiguity) (PerWorkspace (ModifiedLayout (Decoration TabbedDecoration DefaultShrinker) Simplest) (ModifiedLayout SmartSpacing (ModifiedLayout Minimize (ModifiedLayout Maximize (MultiToggle (HCons REFLECTX EOT) (MultiToggle (HCons REFLECTY EOT) (Choose (ModifiedLayout Rename MultiCol) (Choose (ModifiedLayout Rename (Mirror MultiCol)) (Choose Grid Full)))))))))) Window

Custom layoutHook. See XMonad.Config for default layout.

Added:

maximize to be able to "temporarily yank the focused window out of the layout to mostly fill the screen" with xK_backslash (see keys').

minimize to "make it possible to minimize windows and temporarily removing them from the layout until they are restored" with xK_m and shift + xK_m, respectively (see keys').

Ability to "reflect a layout horizontally or vertically" and to toggle those reflections (XMonad.Layout.Reflect and XMonad.Layout.MultiToggle). Key bindings xK_x and xK_y in keys'.

Added lessBorders. Since smartBorders (= lessBorders Never) only removes borders if "there is only one screen and only one window" or "a floating window covers the entire screen", we use lessBorders Screen to never draw borders on singleton screens. To differentiate between screen focus, we could use the status bar in the future. noBorders can be used for specific layouts.

Added smartSpacing to "add a configurable amount of space around windows", "except when the window is the only visible window on the current workspace".

tabTheme :: Theme

Theme used for tabs with tabbed.

layoutExperimental :: ModifiedLayout (Decoration TabbedDecoration DefaultShrinker) Simplest Window

Experimental layout used on last workspace.

manageHook' :: ManageHook

Custom manageHook for managing new windows. E.g. float a window or shift it to a specific workspace.

eventHook' :: Event -> X All

Custom handleEventHook.

Compose event hooks with <+> or a list with composeAll.

Added ewmhDesktopsEventHook to be able to react to

_NET_CURRENT_DESKTOP (switching desktops)

_NET_WM_DESKTOP (move windows to other desktops)

_NET_ACTIVE_WINDOW (activate another window, changing workspace if needed) and fullscreenEventHook to "handle applications that wish to fullscreen using the _NET_WM_STATE protocol".

logHook' :: X ()

Custom logHook.

Added ewmhDesktopsLogHook to "notify pagers and window lists, such as those in the gnome-panel of the current state of workspaces and windows".

startupHook' :: X ()

Custom startupHook.

Added ewmhDesktopsStartup to "initialize EwmhDesktops and advertise EWMH support to the X server".

configDir :: IO FilePath

Directory of xmonad.hs.

haddockDir :: IO FilePath

Directory used for haddock output.

iconRoot :: String

Icon root directory used for status bars.

dmenu :: String

dmenu with command line options.

trayer :: String

trayer with command line options. TODO calc margin TODO height = statusBarHeight - distance

xftFont :: String

The X Free Type Xft font name.

An Xft Tutorial.

xftFontBold :: String

Bold version of xftFont.

defaultPP'

Arguments

:: (String -> String -> String -> String)
color format function

-> (String -> String)
icon format function

-> PP

Default pretty printing options.

ppLayout'

Arguments

:: (String -> String)
icon format function

-> String
layout name

-> String

Icon string depending on layout name.

xmobarWeatherStationId :: String

Weather StationID used with Xmobar's System Monitor Plugin Weather.

xmobarCommands :: [String]

Configuration of Xmobar plugins.

xmobarTemplate :: [String]

Ordered list of Xmobar plugins.

xmobarTemplateSep :: String

Seperator used between Xmobar plugin templates.

xmobar' :: LayoutClass l Window => XConfig l -> IO (XConfig (ModifiedLayout AvoidStruts l))

Xmobar with command line options.

xmobarFocusedPP :: PP

Custom pretty printing options for xmobar on focused screen.

xmobarUnfocusedPP :: PP

Custom pretty printing options for xmobar on unfocused screen.

dzen' :: LayoutClass l Window => XConfig l -> IO (XConfig (ModifiedLayout AvoidStruts l))

Dzen with command line options.

dzenFocusedPP :: PP

Custom pretty printing options for dzen on focused screen.

dzenUnfocusedPP :: PP

Custom pretty printing options for dzen on unfocused screen.

toggleStrutsKey :: XConfig t -> (KeyMask, KeySym)

Key binding to hide status bar.

Same as in XMonad.Hooks.DynamicLog, since it's not exported.

keys' :: [((KeyMask, KeySym), X ())]

List of additional key bindings.

config' :: XConfig (ModifiedLayout Rename (ModifiedLayout (ConfigurableBorder Ambiguity) (PerWorkspace (ModifiedLayout (Decoration TabbedDecoration DefaultShrinker) Simplest) (ModifiedLayout SmartSpacing (ModifiedLayout Minimize (ModifiedLayout Maximize (MultiToggle (HCons REFLECTX EOT) (MultiToggle (HCons REFLECTY EOT) (Choose (ModifiedLayout Rename MultiCol) (Choose (ModifiedLayout Rename (Mirror MultiCol)) (Choose Grid Full)))))))))))

Override default configuration.

main :: IO ()

Main entry point.

Possible status bars: xmobar', dzen'.

When using statusBar, the following is already taken care of:

Status bar spawned via spawnPipe

dynamicLogWithPP added to logHook

ppOutput added to PP

avoidStruts added to layoutHook

manageDocks added to manageHook, docksEventHook to handleEventHook and docksStartupHook to startupHook via docks config modifier.

xmobarIcon :: String -> String

Xmobar icon string with given icon name.

The iconRoot is passed in xmobar'.

Example:

xmobarIcon "myIcon" -> "<icon=myIcon/>"

dzenIcon :: String -> String

Dzen icon string with given icon name.

Example:

dzenIcon "myIcon" -> "^i(iconRoot/myIcon)"

quote :: String -> String

String in quotes for use with command line.

Example:

quote "hello" -> "\"hello\""

xmobarArgs :: [String] -> String

Formatted Xmobar arguments for use with command line.

Example:

xmobarArgs ["-t", "Mem: <usedratio>%"] -> "[\"-t\", \"Mem: <usedratio>%\"]"

dzenEvents

Arguments

:: [(String, [(String, [String])])]
dzen events: [ (event1, [ (action1, [option1, option2, ...]), (action2, [...]), ... ]) , (event2, [...]) , ...]

-> String
formatted events

Formatted Dzen events, actions and options for use with command line.

Example:

events = [ ("button1", [("exec", ["xterm", "firefox"])])
         , ("entertitle", [("uncollapse", []), ("unhide", [])])
         , ("button3", [("exit", [])])
         ]

dzenEvents events -> "button1=exec:xterm:firefox;entertitle=uncollapse,unhide;button3=exit"

listString :: [String] -> String

List of String as String.

Example:

listString ["an", "example"] -> "[an, example]"

pad' :: String -> String -> String

String padded with the same string on both sides.

Example:

pad' "%" "hello" -> "%hello%"

runHaddock :: MonadIO m => m ()

Run haddock on xmonad.hs and the output in a "doc" folder next to it.

The used haddock interfaces are defined in iHaddock.

iHaddock

Arguments

:: [(String, String)]
Documentation base URL paired with path to haddock file.

List of haddock interfaces.